test updatescore function
