+++ package.json
@@ -1,29 +1,42 @@
 {
-  "name": "phaser3-project-template-3",
-  "version": "1.0.0",
-  "description": "A Phaser 3 Project Template",
-  "main": "src/index.js",
-  "scripts": {
-    "build": "webpack",
-    "start": "npm run build && webpack-dev-server --port=8000"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://scottwestover@bitbucket.org/creativerendering/phaser3-template.git"
-  },
-  "license": "MIT",
-  "licenseUrl": "http://www.opensource.org/licenses/mit-license.php",
-  "bugs": {
-    "url": "https://github.com/photonstorm/phaser3-project-template/issues"
-  },
-  "homepage": "https://github.com/photonstorm/phaser3-project-template#readme",
-  "devDependencies": {
-    "webpack-dev-server": "^2.11.0",
-    "webpack": "^3.4.1",
-    "raw-loader": "^0.5.1",
-    "phaser": "^3.3.0"
-  },
-  "dependencies": {
-    "phaser3-rex-plugins": "^1.1.28"
-  }
+	"name": "phaser3-project-template-3",
+	"version": "1.0.0",
+	"description": "A Phaser 3 Project Template",
+	"main": "src/index.js",
+	"scripts": {
+		"test": "jest",
+		"build": "webpack",
+		"start": "npm run build && webpack-dev-server --port=8000"
+	},
+	"repository": {
+		"type": "git",
+		"url": "git+https://scottwestover@bitbucket.org/creativerendering/phaser3-template.git"
+	},
+	"license": "MIT",
+	"licenseUrl": "http://www.opensource.org/licenses/mit-license.php",
+	"bugs": {
+		"url": "https://github.com/photonstorm/phaser3-project-template/issues"
+	},
+	"homepage": "https://github.com/photonstorm/phaser3-project-template#readme",
+	"devDependencies": {
+		"@babel/core": "^7.9.0",
+		"@babel/plugin-proposal-class-properties": "^7.8.3",
+		"@babel/plugin-proposal-export-default-from": "^7.8.3",
+		"@babel/plugin-transform-runtime": "^7.9.0",
+		"@babel/preset-env": "^7.9.0",
+		"@babel/preset-react": "^7.9.4",
+		"@babel/runtime": "^7.9.2",
+		"babel-core": "^6.26.3",
+		"babel-loader": "^8.1.0",
+		"babel-polyfill": "^6.26.0",
+		"babel-preset-env": "^1.7.0",
+		"jest": "^26.4.2",
+		"phaser": "^3.3.0",
+		"raw-loader": "^0.5.1",
+		"webpack": "^3.4.1",
+		"webpack-dev-server": "^2.11.0"
+	},
+	"dependencies": {
+		"phaser3-rex-plugins": "^1.1.28"
+	}
 }
